refactor: move client-server tests to integration tests directory
Use only the public API, add CancellationToken for graceful server
shutdown, and find_available_port helper for test isolation.
